fix : resolve Mateo comments to clean code
