+++ apps/backend/src/queries/project.queries.ts
@@ -1,3 +1,4 @@
+import type { UpdatedAtFilter, UserRole } from '@nao/shared';
 import { and, asc, desc, eq, gt, gte, lte, or, type SQL, sql } from 'drizzle-orm';
 
 import type { AgentSettings, DBProject, DBProjectMember, NewProject, NewProjectMember } from '../db/abstractSchema';
@@ -5,13 +6,7 @@ import s from '../db/abstractSchema';
 import { db } from '../db/db';
 import dbConfig, { Dialect } from '../db/dbConfig';
 import { env } from '../env';
-import type {
-	ListProjectChatsResponse,
-	ProjectChatsFacetKey,
-	UpdatedAtFilter,
-	UserRole,
-	UserWithRole,
-} from '../types/project';
+import type { ListProjectChatsResponse, ProjectChatsFacetKey, UserWithRole } from '../types/project';
 import { HandlerError } from '../utils/error';
 
 export const getProjectByPath = async (path: string): Promise<DBProject | null> => {
