+++ apps/backend/src/types/project.ts
@@ -1,6 +1,4 @@
-export type UserRole = 'admin' | 'user' | 'viewer';
-
-export const USER_ROLES = ['admin', 'user', 'viewer'] as const satisfies readonly UserRole[];
+import type { UserRole } from '@nao/shared';
 
 export interface UserWithRole {
 	id: string;
@@ -26,8 +24,6 @@ export interface ProjectChatListItem {
 	toolErrorCount: number;
 	toolAvailableCount: number;
 }
-
-export type UpdatedAtFilter = { mode: 'single'; value: string } | { mode: 'range'; start: string; end: string };
 
 export interface ListProjectChatsResponse {
 	chats: ProjectChatListItem[];
