+++ apps/frontend/src/lib/utils.ts
@@ -29,3 +29,10 @@ export function capitalize(str: string): string {
 export function isLast<T>(item: T, array: T[]): boolean {
 	return item === array.at(-1);
 }
+
+export function toLocalDateString(d: Date): string {
+	const y = d.getFullYear();
+	const m = String(d.getMonth() + 1).padStart(2, '0');
+	const day = String(d.getDate()).padStart(2, '0');
+	return `${y}-${m}-${day}`;
+}
