+++ apps/shared/src/index.ts
@@ -1,4 +1,5 @@
 export * from './chart-builder';
 export * from './mentions';
 export * from './pluralize';
+export * from './types';
 export * from './utils';
